# ENH: add cyclic GC support to object arrays. Closes #6581.