… numpy/core/tests/test_multiarray.py