There are a few things I'm uncertain about here:

- Currently `ndarray` has custom `tp_alloc` and `tp_free` functions, but they don't seem to do anything different to the defaults. Combining cyclic garbage support with custom `tp_alloc` and `tp_free` functions is apparently difficult and I couldn't get it to work, so I just got rid of them. Is that acceptable? It doesn't cause any harm as far as I can tell.

- I don't really understand `UPDATEIFCOPY`. With the more normal use of `base`, the appropriate thing to do is to visit it in `array_traverse` (because `base` can participate in reference cycles) but ignore it in `array_clear` (because you can't have a reference cycle purely consisting of `base`s, AFAIK). I have a feeling we might need to do something special for arrays with `UPDATEIFCOPY` set.

- the `array_traverse` code is largely copied from `PyArray_INCREF`, and I included the code that deals with discontiguous and misaligned arrays. However, `PyArray_INCREF` and friends need to support arrays that don't own their own data, but `array_traverse` (or at least this section of it) doesn't. Is it possible to have a discontiguous or misaligned object array (or array with object fields) that owns its data? If not, `array_traverse` can be simplified quite a bit.

- there are couple of places in `array_dealloc` where the array's refcount is incremented to prevent numpy API calls from triggering `array_dealloc` again (which would be catastrophic). This is harmless, but it makes it difficult to debug GC stuff, because it makes the total refcount go up every time an object array (or an array with `UPDATEIFCOPY` set) is created and then collected. Is there a straightforward way of avoiding this? I can't think of anything.