@njsmith 2016-11-25
Can you just do

for (int i = 0; i < size; ++i) {
    PyObject* tmp = arr[i];
    arr[i] = NULL;
    DECREF(tmp);
}

? Or is the concern that the DECREF might run arbitrary code, including
writing back to the array's backing store and undoing some of that
clearing? What does PyList do here?

I wouldn't worry too much about UPDATEIFCOPY in cycles -- if an
UPDATEIFCOPY array is in a GC cycle then that's already a bug for sure. The
UPDATEIFCOPY API very much forces the user to make sure that they know
exactly when the UPDATEIFCOPY array is deallocated. There's even a proposal
to deprecate UPDATEIFCOPY in favor of an API that makes this explicit --
it's the last thing needed to get numpy-on-pypy passing the test suite. See
https://mail.scipy.org/pipermail/numpy-discussion/2016-November/076287.html

On Nov 25, 2016 7:23 AM, "J-Sand" <notifications@github.com> wrote:

> OK, I realised my implementation of array_clear isn't quite correct. The
> python docs warn that object pointers have to be set to zero in tp_clear
> before the objects are DECREF'd. I assumed this was in case DECREFing these
> objects ultimately causes the array's own deallocator to fire, so I wrapped
> the DECREFs in Py_INCREF(self); .... Py_DECREF(self); to prevent that.
> However, python's GC code already does that for you. The real problem is
> that the contained objects' deallocators might directly do something to the
> array themselves. One example where this would happen is if an array
> contains an array that is an UPDATEIFCOPY copy of itself. A more
> reasonable example would be a custom C extension type.
>
> So, somehow we need to set data to zero in tp_clear before we start
> DECREFing any of the pointers. As far as I can see, the easiest way of
> doing that is to work out how many objects are in the array, allocate a
> plain C array of that size, copy all of the object pointers into there,
> dealloc and zero data, and then finally walk through the C array and
> DECREF all of the objects. Is there a simpler option? I don't think
> simply copying the ndarray object is an option, in case it's a subclass
> that does something weird when you copy it (it could maybe even create new
> ref cycles).
>
> This also leaves us with the problem of what should happen if the
> deallocator of an array with UPDATEIFCOPY fires and it turns out that
> base->data has already been cleared by its own tp_clear function. Could
> we just document that arrays with UPDATEIFCOPY set are not guaranteed to
> update the base array if they somehow end up in a reference cycle?
>
> An example of an UPDATEIFCOPY ref cycle is as follows (maybe there is a
> simpler way of doing it?):
>
> a = np.zeros(1, dtype=([('a', 'f8'), ('b', 'O')]))
> a['b'][0] = []
> i = np.nditer(a, ['refs_ok'],
>         [['readwrite', 'updateifcopy']],
>         casting='same_kind',
>         op_dtypes=[np.dtype([('a', 'f4'), ('b', 'O')])])
> a['b'][0].append(i.operands[0])
> assert a['b'][0][0].base is a
> assert a['b'][0][0].flags['UPDATEIFCOPY']
>
> Also, I suppose iterator objects are going to need some GC code adding
> too. Are there any other C-implemented types that can participate in ref
> cycles?
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/numpy/numpy/pull/8303#issuecomment-262978384>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAlOaHSF_Zh4T362Ud1EhDt-dJXuB3_lks5rBv1qgaJpZM4K6zXp>
> .
>