@J-Sand 2016-11-25
`list_clear` copies the number of items and the pointer to the first item into temps, sets them both to zero, and then XDECREFs everything (there is a comment pointing out that the XDECREFs could potentially cause the list to grow again). The problem with that approach is that I can't see a straightforward way of using the iteration API to step through the array once we have copied and zeroed `data`.

I think you're right, it would be OK to leave `data` alone at first, but zero the individual object pointers just before you DECREF them (or maybe set them to `Py_None` just in case anything tries to use them without checking for null pointers?). I'll try that approach. Do you know if I can assume that arrays that own their data are contiguous and aligned? I haven't come across a way of creating one that isn't.