@J-Sand 2016-11-29
Thanks for the help!

> It's hard to evaluate "some getting much faster and some much slower" though :-)

I ran the benchmarks with something like `run_tests.py --bench-compare HEAD~2 bench_core`, but now I realise it gives me wildly different results every time. Maybe I need some command-line options to increase the number of loops, or it's intended to be run in environments more stable than my laptop?

I ran some benchmarks with `perf` on `np.empty` with dtypes of `float` and `object` and shapes of `1` and `100`. The impact seems pretty negligible:

    before:
    small_empty_float_array: Median +- std dev: 7.21 us +- 0.30 us
    large_empty_float_array: Median +- std dev: 7.13 us +- 0.32 us
    small_empty_object_array: Median +- std dev: 11.8 us +- 0.5 us
    large_empty_object_array: Median +- std dev: 12.8 us +- 0.7 us

    after:
    small_empty_float_array: Median +- std dev: 7.24 us +- 0.26 us
    large_empty_float_array: Median +- std dev: 7.18 us +- 0.30 us
    small_empty_object_array: Median +- std dev: 11.9 us +- 0.5 us
    large_empty_object_array: Median +- std dev: 13.0 us +- 0.6 us

There really shouldn't be much CPU overhead, because tracking/untracking objects just consists of adding/removing them to a doubly linked list. There is no ongoing overhead for untracked objects, and even for object arrays I don't think the traversal should be hugely expensive.

I think the memory is a bigger problem - allocating 24 extra bytes (at least on 64-bit linux) for every array object, which almost always goes unused, is clearly not ideal.

> I guess we could smuggle this information in using a thread-local... it's ugly, but depending on how much overhead the GC stuff adds it might be worth it.

Hmm... the biggest problem is that `tp_free` needs to know which free function to call. We can just have `tp_free` do a `PyDataType_REFCHK`, but then we need to be able to guarantee that all arrays were allocated using the appropriate allocator, and third-party C extension code might fail to do that (for example, somebody might have a subtype with a custom `tp_new`, or they might build an `ndarray` by calling its `tp_alloc` directly, bypassing our `tp_new`).

I think a new flag would work. We could have `tp_alloc` use the non-GC allocator, and everywhere we call `tp_alloc` we could check whether we want an object array, and if so use the GC allocator instead of `tp_alloc` and set the flag. Then `tp_free` could just look at the flag. If an extension module fails to do this dance in the places where they call `tp_alloc`, they won't get an invalid free, they just won't get cyclic GC support (or we could make the GC allocator the default, and they would waste a bit of memory). And if they already bypass our `tp_alloc` altogether, they really should be providing their own `tp_free`, right?

That's the best option I can think of anyway.